PaaS/Development/API_Integrations/requirements.txt
boto3>=1.26.0
apache-libcloud>=3.7.0
azure-mgmt-compute>=29.1.0
azure-mgmt-resource>=21.1.0
azure-mgmt-network>=21.0.1
azure-identity>=1.12.0
pyyaml>=6.0.1
requests>=2.28.2
python-keystoneclient>=5.1.0
python-novaclient>=18.3.0
libvirt-python>=9.0.0
pytest>=7.4.0
pytest-mock>=3.11.1
pytest-cov>=4.1.0
moto>=4.2.0
responses>=0.23.1
python-dotenv>=1.0.0
